+++ CHANGELOG.md
@@ -16,8 +16,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Changed structure of main() to accommodate optional input data
 - Introduced warning when a RefSeq entry is specified
-- Added support for SMILES, compound synonym names, compound evidence
+- Added support for SMILES, compound synonym names, compound evidence, biological activities, biological targets
 - Specified allowed entries/regular expressions to Base() class for inheritance
+- Moved constants to Base class -> class attributes
+- Enabled multiple references per MIBiG entry
+- Removed support for patent references
+
 
 
 ## [0.3.0] 12-06-2023